test: move apps ducks tests files :truck:
