+++ test/ducks/apps/components/permissionsList.spec.js
@@ -5,10 +5,10 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import { tMock } from '../../jestLib/I18n'
-import { PermissionsList } from '../../../src/ducks/apps/components/PermissionsList'
+import { tMock } from '../../../jestLib/I18n'
+import { PermissionsList } from '../../../../src/ducks/apps/components/PermissionsList'
 
-import mockAppVersion from './_mockAppRegistryVersion'
+import mockAppVersion from '../_mockAppRegistryVersion'
 
 describe('MyApplications component', () => {
   beforeAll(() => {
